backend/requirements.txt
SQLAlchemy==1.4.0b3
sqlalchemy-utils
grpcio
grpcio-health-checking
protobuf
webauthn
pyjwt
eth_keys
eth_account
eth_rlp
hexbytes
web3
requests
pymysql
python-bitcoinlib
sqlalchemy-stubs
pylint
black
pytz
coverage
python-dotenv
soft-webauthn
ipfshttpclient
py-cid
isort
grpc-stubs
grpc_interceptor
../protobufs/python
../common
git+https://github.com/ravirahman/petlib@3b28c97955655671fbc167e007ab6f7ea438801b#egg=petlib
